+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of magerun.
 
+## 2.1.3:
+
+* update install command: check for required php extensions (fixes #1 / thx tatemz)
+* add set_permissions action (fixes #2 / thx tatemz)
+
 ## 2.1.2:
 
 * update install command